[BOOTVID] Move the VGA code into the PC directory. (#2510)
