+++ drivers/base/bootvid/CMakeLists.txt
@@ -3,9 +3,9 @@ spec2def(bootvid.dll bootvid.spec ADD_IMPORTLIB)
 
 if((ARCH STREQUAL "i386") OR (ARCH STREQUAL "amd64"))
     list(APPEND SOURCE
-        i386/bootvid.c
-        i386/bootdata.c
-        i386/vga.c)
+        i386/pc/bootvid.c
+        i386/pc/bootdata.c
+        i386/pc/vga.c)
 elseif(ARCH STREQUAL "arm")
     list(APPEND SOURCE
         arm/bootvid.c)
