+++ drivers/base/bootvid/i386/pc/bootdata.c
