+++ drivers/base/bootvid/i386/pc/bootvid.c
