+++ drivers/base/bootvid/i386/pc/vga.c
