+++ drivers/base/bootvid/i386/pc/vga.h
